test: move `source` cases into subpackage
